JUnit Test: ReadFileTest() and Model.Java  -Aritra Paul
Added unShuffledCodenames to store all the available codenames.
Added JUnit test case for ReadFileTest(), it calls model and the
methods from the model that reads the file then I created a for loop
which converts the ArrayList of String to only one appended String.
This way I could compare the list with the actual GameWords.txt
by hard coding it and not having to write a similar code as
readFiles() method.