SpyMaster: clue and count: updated so clue and count text boxes are functionable,
added methods to determine if clue and count are valid ie. count is a
integer > 0 and clue is not a codename, if invalid; popup indicates
error.

Team:
clue and count are visible, Jbuttons are added for non revealed
locations and are semi-functionable, _menuPanelDisplays whose turn it
is.

I started to seperate some of the larger methods into smaller ones to
improve readability